masfactory/MASFactory-main/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.60.0
anyio==4.10.0
async-timeout==5.0.1 ; python_full_version < '3.11'
attrs==25.4.0
blis==1.3.3
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.13
datasets==4.6.1
diffusers==0.36.0
dill==0.4.0
distro==1.9.0
docstring-parser==0.17.0
exceptiongroup==1.3.0 ; python_full_version < '3.11'
filelock==3.25.0
fire==0.7.1
frozenlist==1.8.0
fsspec==2026.2.0
google-auth==2.40.3
google-genai==1.28.0
h11==0.16.0
hf-xet==1.3.2 ; platform_machine == 'AMD64' or platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'arm64' or platform_machine == 'x86_64'
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.5.0
human-eval==1.0.3
idna==3.10
importlib-metadata==8.7.1
jinja2==3.1.6
jiter==0.10.0
markdown==3.9
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
multidict==6.7.1
multiprocess==0.70.18
murmurhash==1.0.15
numpy==2.2.6 ; python_full_version < '3.11'
numpy==2.3.2 ; python_full_version >= '3.11'
openai==1.98.0
packaging==25.0
pandas==2.3.3 ; python_full_version < '3.11' or python_full_version >= '3.14'
pandas==3.0.1 ; python_full_version >= '3.11' and python_full_version < '3.14'
pillow==12.1.1
preshed==3.0.12
propcache==0.4.1
psutil==7.0.0
pyarrow==23.0.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.7
pydantic-core==2.33.2
pydub==0.25.1
pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1 ; python_full_version < '3.11' or python_full_version >= '3.14'
pyyaml==6.0.3
regex==2026.1.15
requests==2.32.4
rich==14.3.3
rsa==4.9.1
safetensors==0.7.0
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
smart-open==7.5.1
sniffio==1.3.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
tenacity==8.5.0
termcolor==3.3.0
thinc==8.3.10
tiktoken==0.12.0
tqdm==4.67.1
typer==0.24.1
typer-slim==0.24.0
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.3 ; python_full_version < '3.11' or python_full_version >= '3.14' or sys_platform == 'emscripten' or sys_platform == 'win32'
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.3
websockets==15.0.1
wrapt==2.1.1
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
